# openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "CalmKit API",
    "description": "Mental wellness app API with emotion logging, AI coaching, anchor phrases, and guided routines",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": ["http"],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "x-auth-token",
      "in": "header",
      "description": "JWT token for authentication"
    }
  },
  "paths": {
...
    "/api/logs/": {
      "get": {
        "tags": ["Logs"],
        "summary": "Get all user logs",
        "description": "Get all emotion logs with optional date range filtering",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter logs from this date (ISO 8601)",
            "type": "string"
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter logs until this date (ISO 8601)",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": ["Logs"],
        "summary": "Add new emotion log",
        "description": "",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Log data",
            "schema": {
              "type": "object",
              "properties": {
                "trigger": {
                  "type": "string",
                  "example": "Meeting with boss"
                },
                "emotion": {
                  "type": "string",
                  "example": "Anxious"
                },
                "intensity": {
                  "type": "number",
                  "example": 7
                },
                "anchor": {
                  "type": "string",
                  "example": "I am grounded"
                },
                "contributing": {
                  "type": "array",
                  "example": ["Stressed", "Tired"],
                  "items": {
                    "type": "string"
                  }
                },
                "time": {
                  "type": "string",
                  "example": "2024-01-01T12:00:00.000Z"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/logs/{id}": {
      "delete": {
        "tags": ["Logs"],
        "summary": "Delete a log",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Log ID"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/logs/mood-trends/{year}/{month}": {
      "get": {
        "tags": ["Logs"],
        "summary": "Get daily mood trends for a specific month",
        "description": "Get daily logs and mood scores for calendar view",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Year (e.g., 2024)"
          },
          {
            "name": "month",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "Month (1-12)"
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/passphrase/": {
      "put": {
        "tags": ["Passphrase"],
        "summary": "Update passphrase and re-encrypt logs",
        "description": "Trigger server-side re-encryption of all user logs",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Passphrase data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "passcode": {
                  "type": "string",
                  "example": "1234"
                },
                "clientSalt": {
                  "type": "string",
                  "example": "base64string"
                }
              }
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/passphrase/status": {
      "get": {
        "tags": ["Passphrase"],
        "summary": "Get passphrase migration job status",
        "description": "Get the latest passphrase migration job status",
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ]
}